requirements.txt
fastapi[standard]
uvicorn
asyncpg
aiohttp
python-dotenv
celery[redis]
requests
fpdf
pysendpulse